Optimized response handler to reduce latency 10x.
The main bottleneck was each call to flickr's REST API, which takes around
250-300 ms per call (!). Unfortunately, flickr.py is not very careful about
avoiding unnecessary calls, so we were making half a dozen calls per hit.

It turns out that all of the required information is received in the photoset
list response. Some hacks were made to make flickr.py store all of the
relevant information, and make it accessible without making additional
unneccessary API calls.

Furthermore, the association between flickr screenname and user id is
memcached, as it is unlikely to ever change. (While not 100% true, this
assumption is good enough for our purposes.)

This reduces the number of flickr API calls per hit to just one, bringing the
total response time from around 3 seconds to under 300 ms.